requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cuser==0.0.4
Django==2.2
django-appconf==1.0.4
django-bootstrap4==2.1.1
django-environ==0.4.5
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pyserial==3.4
pytz==2020.1
requests==2.24.0
serializers==0.2.4
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
